image_processing/requirements.txt
black==23.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
filelock==3.9.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0rc1
numpy==1.24.2
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycodestyle==2.10.0
requests==2.28.1
sympy==1.11.1
torch==2.1.0.dev20230407
torchaudio==2.1.0.dev20230407
torchvision==0.16.0.dev20230407
typing_extensions==4.4.0
urllib3==1.26.13
